Move SysTool from skip.py to systool.py
This avoids possible import cycle in the generated testfile.
